fix: elements iterator transpose behavior
Transpose was working for 2D but not higher dimensions, so how
elements were iterated was changed to correct the error.

Also, the elements are now copied if a view is taken of a view,
where the first view was a transposed array.

That means those operations take linear time in that particular
case, but this appears to be the same thing that numpy does.

Otherwise, there's no way to know how to iterate through the
elements for sure.