refactor: define strides calculated from shape
There's another possible feature to add here. Right now, when
reshaping is done on a transposed array (such that `contiguous` is
`false`), the elements are copied to a different array because
there's no guarantee that iteration order can be determined after
the reshape is done.

However, that would be possible under certain conditions, which, I
think, include that the new shape the array is being put into is
composed of factors of the current shape. For example, the
following reshape doesn't require copying.

`Seq(100, 100)` -> `Seq(5, 10, 2, 10, 10)`

Can think of it as the first 100 getting split up into 5, 10, and
2. Below are a few examples when copying is necessary:

`Seq(100, 100)` -> `Seq(10000)`
`Seq(100, 100)` -> `Seq(1000, 10)`
`Seq(100, 100)` -> `Seq(25, 10, 4, 10)`